refactor(core/config_models): step 1 - estrarre path_normalization.py
Crea toolkit/core/config_models/path_normalization.py con:
- _err, _require_map
- _resolve_path_value, _SECTION_PATH_WHITELIST
- _path_tokens_to_str, _set_nested_value
- _iter_matching_tokens, _get_nested_value
- _normalize_section_paths
- _is_managed_output_root, _resolve_root, _ensure_root_within_repo

common.py ora importa da path_normalization e ri-esporta per backward
compat via __init__.py (tutti i consumer esistenti continuano a funzionare
importando da common, non da path_normalization direttamente).

63/63 test passati, ruff all green.
